# Write Nav links when metadata full