--- a/test/Microsoft.AspNetCore.OData.Tests/Formatter/Serialization/ODataResourceSerializerTests.cs
+++ b/test/Microsoft.AspNetCore.OData.Tests/Formatter/Serialization/ODataResourceSerializerTests.cs
@@ -322,8 +322,11 @@ namespace Microsoft.AspNetCore.OData.Tests.Formatter.Serialization
             serializer.Setup(s => s.CreateNavigationLink(selectExpandNode.SelectedNavigationProperties.ElementAt(0), It.IsAny<ResourceContext>())).Verifiable();
             serializer.Setup(s => s.CreateNavigationLink(selectExpandNode.SelectedNavigationProperties.ElementAt(1), It.IsAny<ResourceContext>())).Verifiable();
 
+            ODataSerializerContext writeContext = new ODataSerializerContext() { NavigationSource = _customerSet, Model = _model, Path = _path };
+            writeContext.MetadataLevel = ODataMetadataLevel.Full;
+
             // Act
-            await serializer.Object.WriteObjectInlineAsync(_customer, _customerType, writer.Object, _writeContext);
+            await serializer.Object.WriteObjectInlineAsync(_customer, _customerType, writer.Object, writeContext);
 
             // Assert
             serializer.Verify();
@@ -360,8 +363,11 @@ namespace Microsoft.AspNetCore.OData.Tests.Formatter.Serialization
             writer.Setup(w => w.WriteStartAsync(navigationLinks[0])).Verifiable();
             writer.Setup(w => w.WriteStartAsync(navigationLinks[1])).Verifiable();
 
+            ODataSerializerContext writeContext = new ODataSerializerContext() { NavigationSource = _customerSet, Model = _model, Path = _path };
+            writeContext.MetadataLevel = ODataMetadataLevel.Full;
+
             // Act
-            await serializer.Object.WriteObjectInlineAsync(_customer, _customerType, writer.Object, _writeContext);
+            await serializer.Object.WriteObjectInlineAsync(_customer, _customerType, writer.Object, writeContext);
 
             // Assert
             writer.Verify();