@habbes 2023-10-24
@ElizabethOkerio would be great to also show CPU improvements if that's available